# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Pokemon API
  version: "1.0.0"
  summary: A PokeAPI-compatible Pokemon database with extra game-specific data.
  description: |
    A read-only HTTP API serving Pokemon data — species, moves, abilities, items,
    encounters, evolution chains, raids, trainers, events, and more. Most
    endpoints are compatible with the public [PokeAPI](https://pokeapi.co/)
    schema; additional endpoints expose data scraped from
    [pokemondb.net](https://pokemondb.net) and other sources, including raid
    rotations, trainer teams, mass outbreaks, version exclusives, in-game
    trades, and Z-moves.

    **Identifiers.** Endpoints that take an `{identifier}` accept either the
    Pokemon's lowercase name (e.g. `pikachu`) or its National Pokedex number
    (e.g. `25`).

    **Pagination.** List endpoints accept `limit` and `offset` query parameters
    and return `{ count, results }` envelopes mirroring PokeAPI.
servers:
  - url: https://poke-api.duocore.dev:158
    description: Public API (HTTPS, port 158 — Totodile's National Dex number).
...
components:
  parameters:
    Limit:
      name: limit
      in: query
      description: Page size. Defaults vary per endpoint (typically 1025 or 20).
      required: false
      schema:
        type: integer
        minimum: 1
    Offset:
      name: offset
      in: query
      description: Number of items to skip.
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    Identifier:
      name: identifier
      in: path
      required: true
      description: Pokemon name (lowercase) or National Dex number.
      schema:
        type: string
      examples:
        byName:
          value: totodile
        byNumber:
          value: "158"
...
  schemas:
...
    PageEnvelope:
      type: object
      description: PokeAPI-style paginated list envelope.
      properties:
        count:
          type: integer
        results:
          type: array
          items:
            type: object
            properties:
              name: { type: string }
              url:  { type: string }
...
paths:
  /api/v2/pokemon:
    get:
      tags: [Pokemon]
      summary: List Pokemon
      description: PokeAPI-compatible paginated list. `limit` defaults to 1025 (all Pokemon).
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Paginated list.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PageEnvelope' }
  /api/v2/pokemon/competitive:
    get:
      tags: [Competitive]
      summary: List competitive-format Pokemon
      description: Returns Pokemon with stats/abilities optimized for competitive use.
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Competitive Pokemon list.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CompetitivePokemon' }
...
  /api/v2/pokemon/{identifier}:
    get:
      tags: [Pokemon]
      summary: Get a Pokemon
      description: PokeAPI-compatible Pokemon object including types, abilities, stats, sprites.
      parameters:
        - $ref: '#/components/parameters/Identifier'
      responses:
        '200':
          description: PokeAPI-shaped Pokemon object.
          content:
            application/json:
              schema:
                type: object
                description: PokeAPI-compatible Pokemon document.
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v2/pokemon/{identifier}/encounters:
    get:
      tags: [Locations & Encounters]
      summary: Get a Pokemon's encounter locations
      parameters:
        - $ref: '#/components/parameters/Identifier'
      responses:
        '200':
          description: PokeAPI-compatible encounter list.
          content:
            application/json:
              schema:
                type: array
                items: { type: object }
  /api/v2/pokemon/{identifier}/moves:
    get:
      tags: [Moves]
      summary: Get a Pokemon's learnable moves
      parameters:
        - $ref: '#/components/parameters/Identifier'
        - name: version_group
          in: query
          schema: { type: string }
          description: Optional version-group slug (e.g. `red-blue`, `scarlet-violet`).
      responses:
        '200':
          description: Moves grouped by learn method.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Move' }
  /api/v2/pokemon/{identifier}/type-defenses:
    get:
      tags: [Types]
      summary: Get type damage multipliers against this Pokemon
      parameters:
        - $ref: '#/components/parameters/Identifier'
      responses:
        '200':
          description: Multipliers per attacking type.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TypeDefense' }
  /api/v2/pokemon-species/{identifier}:
    get:
      tags: [Species & Evolution]
      summary: Get a Pokemon species
      description: PokeAPI-compatible species document — flavor text, color, habitat, capture rate, evolution chain ref.
      parameters:
        - $ref: '#/components/parameters/Identifier'
      responses:
        '200':
          description: Species document.
          content:
            application/json:
              schema: { type: object }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v2/evolution-chain/{id}:
    get:
      tags: [Species & Evolution]
      summary: Get an evolution chain
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
          description: Evolution chain ID (matches species ID at the chain root).
      responses:
        '200':
          description: PokeAPI-compatible evolution chain.
          content:
            application/json:
              schema: { type: object }
...
  /api/v2/type/{identifier}:
    get:
      tags: [Types]
      summary: Get a type
      description: Damage relations, moves, Pokemon of this type.
      parameters:
        - $ref: '#/components/parameters/Identifier'
      responses:
        '200':
          description: Type document.
          content:
            application/json:
              schema: { type: object }
  /api/v2/type/{identifier}/names:
    get:
      tags: [Types]
      summary: Localized names for a type
      parameters:
        - $ref: '#/components/parameters/Identifier'
      responses:
        '200':
          description: Localized name list.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PokemonName' }
  /api/v2/item/{identifier}/names:
    get:
      tags: [Items & Berries]
      summary: Localized names for an item
      parameters:
        - $ref: '#/components/parameters/Identifier'
      responses:
        '200':
          description: Localized name list.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PokemonName' }
  /api/v2/move/{identifier}/names:
    get:
      tags: [Moves]
      summary: Localized names for a move
      parameters:
        - $ref: '#/components/parameters/Identifier'
      responses:
        '200':
          description: Localized name list.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PokemonName' }
  /api/v2/ability/{identifier}/names:
    get:
      tags: [Abilities]
      summary: Localized names for an ability
      parameters:
        - $ref: '#/components/parameters/Identifier'
      responses:
        '200':
          description: Localized name list.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PokemonName' }
  /api/v2/generation/{id}:
    get:
      tags: [Pokedex & Games]
      summary: Get a generation
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Generation document.
          content:
            application/json:
              schema: { type: object }
...
  /api/v2/pokemon-color/{color}:
    get:
      tags: [Egg Groups & Classification]
      summary: List Pokemon by color
      parameters:
        - name: color
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Pokemon of the given color.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PokemonListEntry' }
...
  /api/v2/pokemon-shape/{shape}:
    get:
      tags: [Egg Groups & Classification]
      summary: List Pokemon by shape
      parameters:
        - name: shape
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Pokemon of the given shape.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PokemonListEntry' }
...
  /api/v2/pokemon-habitat/{habitat}:
    get:
      tags: [Egg Groups & Classification]
      summary: List Pokemon by habitat
      parameters:
        - name: habitat
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Pokemon of the given habitat.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PokemonListEntry' }
...
  /api/v2/egg-group/{name}:
    get:
      tags: [Egg Groups & Classification]
      summary: List Pokemon in an egg group
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Pokemon in the egg group.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PokemonListEntry' }
  /api/v2/version-group/{name}:
    get:
      tags: [Pokedex & Games]
      summary: Get a version group
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
          description: Slug like `red-blue`, `gold-silver`, `scarlet-violet`.
      responses:
        '200':
          description: Version group document.
          content:
            application/json:
              schema: { type: object }
  /api/v2/pokedex/{identifier}:
    get:
      tags: [Pokedex & Games]
      summary: Get a Pokedex
      description: |
        PokeAPI-compatible regional Pokedex. Pass `?game=<slug>` to scope flavor
        entries to a specific game.
      parameters:
        - $ref: '#/components/parameters/Identifier'
        - $ref: '#/components/parameters/GameQuery'
      responses:
        '200':
          description: Pokedex document.
          content:
            application/json:
              schema: { type: object }
  /api/v2/pokedex/game/{game}:
    get:
      tags: [Pokedex & Games]
      summary: Get a game's Pokedex
      parameters:
        - name: game
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Regional Pokedex entries.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/RegionalDexEntry' }
...
  /api/v2/item:
    get:
      tags: [Items & Berries]
      summary: List items
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Items list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: { type: integer }
                  results:
                    type: array
                    items: { $ref: '#/components/schemas/ItemDetail' }
  /api/v2/item/{identifier}:
    get:
      tags: [Items & Berries]
      summary: Get an item
      parameters:
        - $ref: '#/components/parameters/Identifier'
      responses:
        '200':
          description: Item details.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ItemDetail' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v2/item/{identifier}/locations:
    get:
      tags: [Items & Berries]
      summary: Where an item can be obtained
      parameters:
        - $ref: '#/components/parameters/Identifier'
        - $ref: '#/components/parameters/GameQuery'
      responses:
        '200':
          description: Item locations.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ItemLocation' }
  /api/v2/move:
    get:
      tags: [Moves]
      summary: List moves
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: generation
          in: query
          schema: { type: integer }
          description: Filter by generation introduced.
        - name: type
          in: query
          schema: { type: string }
          description: Filter by move type (e.g. `water`, `fire`).
        - name: category
          in: query
          schema: { type: string }
          description: Filter by category (`physical`, `special`, `status`).
      responses:
        '200':
          description: Move list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: { type: integer }
                  results:
                    type: array
                    items: { $ref: '#/components/schemas/MoveDetail' }
  /api/v2/move/{identifier}:
    get:
      tags: [Moves]
      summary: Get a move
      parameters:
        - $ref: '#/components/parameters/Identifier'
      responses:
        '200':
          description: Move details.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MoveDetail' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v2/ability:
    get:
      tags: [Abilities]
      summary: List abilities
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Ability list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: { type: integer }
                  results:
                    type: array
                    items: { $ref: '#/components/schemas/AbilityDetail' }
  /api/v2/ability/{identifier}:
    get:
      tags: [Abilities]
      summary: Get an ability
      description: Returns ability details and the Pokemon that have it.
      parameters:
        - $ref: '#/components/parameters/Identifier'
      responses:
        '200':
          description: Ability with Pokemon list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  ability: { $ref: '#/components/schemas/AbilityDetail' }
                  pokemon:
                    type: array
                    items: { $ref: '#/components/schemas/AbilityPokemon' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v2/berry/{identifier}:
    get:
      tags: [Items & Berries]
      summary: Get a berry
      parameters:
        - $ref: '#/components/parameters/Identifier'
      responses:
        '200':
          description: Berry details.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Berry' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v2/pokemon/{identifier}/flavor-text:
    get:
      tags: [Species & Evolution]
      summary: Pokedex flavor text per game
      parameters:
        - $ref: '#/components/parameters/Identifier'
      responses:
        '200':
          description: Flavor text entries.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PokedexEntry' }
  /api/v2/pokemon/{identifier}/names:
    get:
      tags: [Pokemon]
      summary: Localized Pokemon names
      parameters:
        - $ref: '#/components/parameters/Identifier'
      responses:
        '200':
          description: Localized names.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PokemonName' }
  /api/v2/pokemon/{identifier}/sprites-all:
    get:
      tags: [Pokemon]
      summary: All sprite URLs for a Pokemon
      parameters:
        - $ref: '#/components/parameters/Identifier'
      responses:
        '200':
          description: Sprite list across generations.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PokemonSprite' }
  /api/v2/pokemon/{identifier}/held-items:
    get:
      tags: [Items & Berries]
      summary: Items wild Pokemon may be holding
      parameters:
        - $ref: '#/components/parameters/Identifier'
      responses:
        '200':
          description: Wild held-item entries.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/WildHeldItem' }
  /api/v2/pokemon/{identifier}/biology:
    get:
      tags: [Pokemon]
      summary: Pokemon biology / dex description
      parameters:
        - $ref: '#/components/parameters/Identifier'
      responses:
        '200':
          description: Biology blurb.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PokemonBiology' }
  /api/v2/pokemon/{identifier}/game-locations:
    get:
      tags: [Locations & Encounters]
      summary: Per-game encounter locations for a Pokemon
      parameters:
        - $ref: '#/components/parameters/Identifier'
      responses:
        '200':
          description: Game-location entries.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PokemonGameLocation' }
  /api/v2/pokemon/{identifier}/forms:
    get:
      tags: [Forms & Variants]
      summary: Alternate forms for a Pokemon
      parameters:
        - $ref: '#/components/parameters/Identifier'
      responses:
        '200':
          description: Form list.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PokemonForm' }
  /api/v2/pokemon/{identifier}/egg-move-parents:
    get:
      tags: [Moves]
      summary: Parents that can pass egg moves to this Pokemon
      parameters:
        - $ref: '#/components/parameters/Identifier'
      responses:
        '200':
          description: Egg move parent list.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/EggMoveParent' }
  /api/v2/tm:
    get:
      tags: [Items & Berries]
      summary: List TM/HM locations
      parameters:
        - $ref: '#/components/parameters/GameQuery'
      responses:
        '200':
          description: TM locations.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TmLocation' }
...
  /api/v2/item-locations:
    get:
      tags: [Items & Berries]
      summary: Browse item locations by game
      parameters:
        - $ref: '#/components/parameters/GameQuery'
      responses:
        '200':
          description: Item locations.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ItemLocation' }
...
  /api/v2/location/pokemon/{name}:
    get:
      tags: [Locations & Encounters]
      summary: Encounter locations for a Pokemon
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/GameQuery'
      responses:
        '200':
          description: Encounter list.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/LocationEncounter' }
  /api/v2/location/region/{region}/routes:
    get:
      tags: [Locations & Encounters]
      summary: List routes in a region
      parameters:
        - name: region
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/GameQuery'
      responses:
        '200':
          description: Route list.
          content:
            application/json:
              schema:
                type: array
                items: { type: string }
  /api/v2/location/region/{region}/route/{route}:
    get:
      tags: [Locations & Encounters]
      summary: Encounters on a specific route
      parameters:
        - name: region
          in: path
          required: true
          schema: { type: string }
        - name: route
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/GameQuery'
      responses:
        '200':
          description: Route encounter list.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/LocationEncounter' }
...
  /api/v2/news/raids/{pokemon}:
    get:
      tags: [Events & News]
      summary: Raid events featuring a specific Pokemon
      parameters:
        - name: pokemon
          in: path
          required: true
          schema: { type: string }
        - name: tera_type
          in: query
          schema: { type: string }
          description: Optional tera type filter.
      responses:
        '200':
          description: Matching raid events plus suggested counters.
          content:
            application/json:
              schema: { type: object }
  /api/v2/pokemon/{identifier}/classification:
    get:
      tags: [Egg Groups & Classification]
      summary: Legendary / mythical / paradox flags for a Pokemon
      parameters:
        - $ref: '#/components/parameters/Identifier'
      responses:
        '200':
          description: Classification record.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PokemonClassification' }
  /api/v2/pokemon/{identifier}/go:
    get:
      tags: [Pokemon]
      summary: Pokemon GO stats
      parameters:
        - $ref: '#/components/parameters/Identifier'
      responses:
        '200':
          description: GO stat record.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PokemonGo' }
  /api/v2/pokemon/{identifier}/contest-stats:
    get:
      tags: [Pokemon]
      summary: Contest stats for a Pokemon
      parameters:
        - $ref: '#/components/parameters/Identifier'
      responses:
        '200':
          description: Per-contest-type appeal/jam values.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ContestStat' }
...
  /api/v2/z-move/{name}:
    get:
      tags: [Z-Moves]
      summary: Get a Z-Move
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Z-Move details.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ZMove' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v2/in-game-trades:
    get:
      tags: [Trades, Tutors & Exclusives]
      summary: List in-game trades
      parameters:
        - $ref: '#/components/parameters/GameQuery'
      responses:
        '200':
          description: In-game trade list.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/InGameTrade' }
  /api/v2/event-pokemon:
    get:
      tags: [Events & News]
      summary: List event-distributed Pokemon
      parameters:
        - name: name
          in: query
          schema: { type: string }
        - $ref: '#/components/parameters/GameQuery'
      responses:
        '200':
          description: Event Pokemon distributions.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/EventPokemon' }
  /api/v2/mass-outbreaks:
    get:
      tags: [Events & News]
      summary: List mass outbreaks (Legends Arceus / SV)
      parameters:
        - $ref: '#/components/parameters/GameQuery'
        - name: region
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Outbreak list.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/MassOutbreak' }
  /api/v2/battle-facility:
    get:
      tags: [Trainers & Battle Facilities]
      summary: List battle facilities
      parameters:
        - $ref: '#/components/parameters/GameQuery'
      responses:
        '200':
          description: Battle facility list.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/BattleFacility' }
  /api/v2/move-tutor:
    get:
      tags: [Trades, Tutors & Exclusives]
      summary: List move tutor locations
      parameters:
        - $ref: '#/components/parameters/GameQuery'
        - name: move
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Move-tutor list.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/MoveTutorLocation' }
...